Update glitch assets system
Previously when run on glitch.com enviroment urls to assets in glitch
CDN were transfermed from relative urls to the absolute CDN one's by a
proxy at the server level and a tranform function at client side. Since
choice-assist is designed to compile into a static set of files which
can be served by any file server, we don't have the option of proxying
urls for glitch assets at the server level when run in production. It
can therefore be confusing to have that happen in development but have
that stop working when the env mode is set to production. So let's just
remove the server side proxy and require all CDN tranformation to be
required on the client side to work.